move function to middleware makes more sense
